Implemented a CommandHandler class, AB#83
